Remove Float and Half Scalar Types (#537)
Remove half scalar type.
Remove Float Scalar Type.

+++ torch/csrc/jit/codegen/cuda/parser.cpp
@@ -402,10 +402,10 @@ class IrParser {
             // TODO: we need to get a proper lower bound per dtype in operand.
             auto low = value_map.count(node->inputs()[1]->unique()) != 0
                 ? value_map[node->inputs()[1]->unique()]
-                : new Float(std::numeric_limits<float>::min());
+                : new Double(std::numeric_limits<float>::min());
             auto high = value_map.count(node->inputs()[2]->unique()) != 0
                 ? value_map[node->inputs()[2]->unique()]
-                : new Float(std::numeric_limits<float>::max());
+                : new Double(std::numeric_limits<float>::max());
 
             auto out = clamp(operand, low, high);
             value_map.emplace(node->output()->unique(), out);
@@ -543,9 +543,9 @@ class IrParser {
             auto x_sum_bcast = broadcast(x_sum, broadcast_mask);
             auto x_mean = div(x_sum_bcast, num_features);
 
-            // auto current_mean_hat = mul(x_mean, new Float(kMomentum));
+            // auto current_mean_hat = mul(x_mean, new Double(kMomentum));
             // auto rmean_bcast = broadcast(running_mean, broadcast_mask);
-            // auto mean_hat = mul(rmean_bcast, new Float(1.0 - kMomentum));
+            // auto mean_hat = mul(rmean_bcast, new Double(1.0 - kMomentum));
             // auto new_mean_hat = add(mean_hat, current_mean_hat);
 
             auto x_mean_sub = sub(input, x_mean);
@@ -556,12 +556,12 @@ class IrParser {
 
             // auto num_feature_decrement = sub(num_features, new Int(1));
             // auto unbiased_var = div(var_sum_bcast, num_feature_decrement);
-            // auto current_var_hat = mul(unbiased_var, new Float(kMomentum));
+            // auto current_var_hat = mul(unbiased_var, new Double(kMomentum));
             // auto rvar_bcast = broadcast(running_var, broadcast_mask);
-            // auto var_hat = mul(rvar_bcast, new Float(1.0 - kMomentum));
+            // auto var_hat = mul(rvar_bcast, new Double(1.0 - kMomentum));
             // auto new_var_hat = add(var_hat, current_var_hat);
 
-            auto var_eps = add(var, new Float(kEps));
+            auto var_eps = add(var, new Double(kEps));
             auto rvar = unaryOp(UnaryOpType::Rsqrt, var_eps);
             auto output = mul(x_mean_sub, rvar);
 
@@ -635,7 +635,7 @@ class IrParser {
             auto var_sum = sum(x_mean_sub_pow, reduction_axes);
             auto var_sum_bcast = broadcast(var_sum, broadcast_mask);
             auto var = div(var_sum_bcast, num_features);
-            auto var_eps = add(var, new Float(kEps));
+            auto var_eps = add(var, new Double(kEps));
             auto rvar = unaryOp(UnaryOpType::Rsqrt, var_eps);
             auto output = mul(x_mean_sub, rvar);
 
@@ -800,17 +800,17 @@ class IrParser {
   bool registerScalar(const JitValue* val) {
     if (val->type()->isSubtypeOf(static_cast<c10::TypePtr>(FloatType::get()))) {
       CgValue cg_val;
-      if (auto ival = constant_as<float>(val)) {
-        cg_val = new Float(ival.value());
+      if (auto ival = constant_as<double>(val)) {
+        cg_val = new Double(ival.value());
       } else {
-        cg_val = new Float();
+        cg_val = new Double();
       }
       value_map_.emplace(val->unique(), cg_val);
       return true;
     } else if (val->type()->isSubtypeOf(
                    static_cast<c10::TypePtr>(IntType::get()))) {
       CgValue cg_val;
-      if (auto ival = constant_as<int>(val)) {
+      if (auto ival = constant_as<int64_t>(val)) {
         cg_val = new Int(ival.value());
       } else {
         cg_val = new Int();
